- Update : Change type from Double to Int, and Create a dedicated Mapper file.

+++ app/src/main/java/com/aetherized/smartfinance/core/database/entity/GoalEntity.kt
@@ -12,9 +12,9 @@ data class GoalEntity(
     val id: Long = 0L,
     val name: String,
     @ColumnInfo(name = "target_amount")
-    val targetAmount: Double,
+    val targetAmount: Int,
     @ColumnInfo(name = "current_amount")
-    val currentAmount: Double = 0.0,
+    val currentAmount: Int = 0,
     val deadline: Long,
     @ColumnInfo(name = "is_achieved")
     val isAchieved: Boolean = false
